+++ ReleaseNotes.md
@@ -5,6 +5,8 @@ Date: 2024.06.02
   - Remove cache folder when finished
   - Remove audio from yt video
   - Added install and start scripts
+  - Added Docker support
+  - Added sheet music generation
 
 # Version: 0.0.10
 Date: 2024.05.01